apps/ai_chatbot_streamlit/requirements.txt
streamlit>=1.32.0
databricks-sdk>=0.23.0
